--- src/components/hooks/index.tsx
@@ -1,13 +1,0 @@
-import * as React from "react";
-import StateHook from "./stateHook";
-
-function HooksHome() {
-  return (
-    <div>
-      <h1>React Hooks</h1>
-      <StateHook />
-    </div>
-  );
-}
-
-export default HooksHome;
