+++ src/index.tsx
@@ -1,14 +1,7 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App";
-import * as serviceWorker from "./serviceWorker";
-import { client } from "./middleware";
-import { ApolloProvider } from "react-apollo";
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import App from "./App";
+import registerServiceWorker from "./registerServiceWorker";
 
-ReactDOM.render(
-  <ApolloProvider client={client}>
-    <App />
-  </ApolloProvider>,
-  document.getElementById("root")
-);
-serviceWorker.unregister();
+ReactDOM.render(<App />, document.getElementById("root") as HTMLElement);
+registerServiceWorker();
