--- src/middleware/index.ts
@@ -1,5 +1,0 @@
-import ApolloClient from "apollo-boost";
-
-export const client = new ApolloClient({
-  uri: "http://localhost:3000/graphiql"
-});
